Restruct project
